+++ travel-planner/src/routes/index.tsx
@@ -4,6 +4,7 @@ import Flights from "../pages/Flights";
 import FlightsResults from "../pages/Flights/results";
 import FlightBooking from "../pages/Flights/booking";
 import FlightItinerary from "../pages/Flights/itinerary";
+import Destination from "../pages/Attractions";
 
 const AppRoutes = () => (
   <Routes>
@@ -11,6 +12,7 @@ const AppRoutes = () => (
     <Route path="/flights/results" element={<FlightsResults />} />
     <Route path="/flight/booking/:id" element={<FlightBooking />} />
     <Route path="/flight/itinerary/:id" element={<FlightItinerary />} />
+    <Route path="/attractions" element={<Destination />} />
   </Routes>
 );